+++ Sources/SpaceEngineers.ObjectBuilders/ObjectBuilders/MyObjectBuilder_AutopilotBase.cs
@@ -7,6 +7,7 @@ namespace Sandbox.Common.ObjectBuilders
     [ProtoContract]
     [MyObjectBuilderDefinition]
     [XmlSerializerAssembly("SpaceEngineers.ObjectBuilders.XmlSerializers")]
+    [XmlInclude(typeof(MyObjectBuilder_SimpleAutopilot))]
     public class MyObjectBuilder_AutopilotBase : MyObjectBuilder_Base
     {
     }